requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
python-magic==0.4.27
httpx==0.28.1
python-jose[cryptography]==3.3.0
jinja2==3.1.4
python-dotenv==1.0.1
langextract==0.1.10
